requirements.txt
opencv-python
tkintertable
numpy
Pillow
